Modularize test
