+++ auth/test_auth_user.py
@@ -97,10 +97,7 @@ class UserAuthTest(django.test.TestCase):
         response = self.client.post(vote_url, form_data)
         # should be redirected to the login page
         self.assertEqual(response.status_code, 302)  # could be 303
-        # TODO: this test fails because reverse('login') does not include
-        # the query parameter ?next=/polls/1/vote/
-        #self.assertRedirects(response, reverse('login') )
-        # How to fix it? 
+        
         login_with_next = f"{reverse('login')}?next={vote_url}"
         self.assertRedirects(response, login_with_next )
